Move getConfig module
