+++ app/callbacks/lib/callbacks.js
@@ -11,7 +11,7 @@ let logger = {
 };
 
 if (Meteor.isClient) {
-	const { getConfig } = require('../../ui-utils/client/config');
+	const { getConfig } = require('../../../client/lib/utils/getConfig');
 	timed = [getConfig('debug'), getConfig('timed-callbacks')].includes('true');
 }
