+++ client/lib/utils/getConfig.ts
@@ -1,0 +1,6 @@
+import { Meteor } from 'meteor/meteor';
+
+export const getConfig = (key: string): string | null => {
+	const searchParams = new URLSearchParams(window.location.search);
+	return searchParams.get(key) || Meteor._localStorage.getItem(`rc-config-${key}`);
+};
