+++ client/views/room/contextualBar/RoomFiles/hooks/useFilesList.ts
@@ -1,12 +1,12 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { getConfig } from '../../../../../../app/ui-utils/client/config';
 import { useEndpoint } from '../../../../../contexts/ServerContext';
 import { useUserRoom, useUserId } from '../../../../../contexts/UserContext';
 import { useScrollableMessageList } from '../../../../../hooks/lists/useScrollableMessageList';
 import { useStreamUpdatesForMessageList } from '../../../../../hooks/lists/useStreamUpdatesForMessageList';
 import { useComponentDidUpdate } from '../../../../../hooks/useComponentDidUpdate';
 import { FilesList, FilesListOptions } from '../../../../../lib/lists/FilesList';
+import { getConfig } from '../../../../../lib/utils/getConfig';
 
 export const useFilesList = (
 	options: FilesListOptions,
