+++ client/views/teams/contextualBar/channels/hooks/useTeamsChannelList.ts
@@ -1,11 +1,11 @@
 import { useCallback, useMemo, useState } from 'react';
 
-import { getConfig } from '../../../../../../app/ui-utils/client/config';
 import { IRoom } from '../../../../../../definition/IRoom';
 import { useEndpoint } from '../../../../../contexts/ServerContext';
 import { useScrollableRecordList } from '../../../../../hooks/lists/useScrollableRecordList';
 import { useComponentDidUpdate } from '../../../../../hooks/useComponentDidUpdate';
 import { RecordList } from '../../../../../lib/lists/RecordList';
+import { getConfig } from '../../../../../lib/utils/getConfig';
 import { mapMessageFromApi } from '../../../../../lib/utils/mapMessageFromApi';
 
 type TeamsChannelListOptions = {
